+++ player.h
@@ -13,12 +13,12 @@ using namespace std;
 using namespace world;
 namespace player {
 	extern Vector3 pos;  // 玩家坐标
-	extern GLfloat vx, vy, vz;  // 速度
-	extern GLfloat vf, vr;  // 正向速度，右向速度
+	extern GLdouble vx, vy, vz;  // 速度
+	extern GLdouble vf, vr;  // 正向速度，右向速度
 	extern Vector2 rot;  // 视角旋转
 	extern bool grounded;  // 是否落地
 	extern GLfloat xSensitivity, ySensitivity;  // 鼠标灵敏度
-	extern GLfloat speed;  // 行走速度
+	extern GLdouble speed;  // 行走速度
 	extern int hWidth, hHeight;  // 屏幕宽高一半
 	extern float val;
 	extern physics::RayCastHit *hit;
